+++ samples/contacts/src/main/java/sample/contact/ContactDao.java
@@ -1,4 +1,4 @@
-/* Copyright 2004 Acegi Technology Pty Limited
+/* Copyright 2004, 2005 Acegi Technology Pty Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,11 +27,11 @@ import java.util.List;
 public interface ContactDao {
     //~ Methods ================================================================
 
-    public Contact getById(Integer id);
+    public Contact getById(Long id);
 
     public void create(Contact contact);
 
-    public void delete(Integer contactId);
+    public void delete(Long contactId);
 
     public List findAll();
