+++ samples/contacts/src/main/java/sample/contact/ContactManager.java
@@ -1,4 +1,4 @@
-/* Copyright 2004 Acegi Technology Pty Limited
+/* Copyright 2004, 2005 Acegi Technology Pty Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ public interface ContactManager {
 
     public List getAllRecipients();
 
-    public Contact getById(Integer id);
+    public Contact getById(Long id);
 
     public Contact getRandomContact();
